requirements.txt
altgraph==0.17.5
anyio==4.12.1
asgiref==3.11.1
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
fonttools==4.61.1
h11==0.16.0
htmltools==0.6.0
idna==3.11
kiwisolver==1.4.9
linkify-it-py==2.0.3
macholib==1.16.4
markdown-it-py==4.0.0
matplotlib==3.10.8
mdit-py-plugins==0.5.0
mdurl==0.1.2
narwhals==2.16.0
numpy==2.4.2
orjson==3.11.7
packaging==26.0
pandas==3.0.0
pillow==12.1.1
platformdirs==4.5.1
prompt_toolkit==3.0.52
pyinstaller==6.18.0
pyinstaller-hooks-contrib==2026.0
pyparsing==3.3.2
PySide6==6.10.2
PySide6_Addons==6.10.2
PySide6_Essentials==6.10.2
python-dateutil==2.9.0.post0
python-multipart==0.0.22
PyYAML==6.0.3
questionary==2.1.1
setuptools==82.0.0
shiboken6==6.10.2
shiny==1.5.1
shinychat==0.2.9
six==1.17.0
starlette==0.52.1
typing_extensions==4.15.0
uc-micro-py==1.0.3
uvicorn==0.40.0
watchfiles==1.1.1
wcwidth==0.6.0
websockets==16.0
